Refactor : 에러 바운더리를 위한 로직 분리
