+++ src/app/(content-layout)/[groupId]/tasklist/_tasklist/components/TasksWiseTask.tsx
@@ -7,7 +7,7 @@ import RemoveTaskModal from './ModalContents/RemoveTaskModal';
 import { useTaskActions } from '../hooks/use-task-actions';
 import { useTaskModals } from '../hooks/use-task-modals';
 import ManageTaskItemModal from './manage-task-item-modal/MangeTaskItemModal';
-import DetailTask from '../../@detailTask/page';
+import DetailTaskContainer from '../../@detailTask/page';
 
 interface Props {
   task: Task;
@@ -71,7 +71,7 @@ export default function TasksWiseTask({ task, groupId, taskListId }: Props) {
             date={safeFormatDate(task.date)}
             frequency={task.frequency}
           />
-          <DetailTask
+          <DetailTaskContainer
             isDone={isDone}
             setIsDone={toggleTaskStatus}
             taskId={task.id}
